fix: fix ext component
