+++ packages/kit/src/views/Setting/pages/List/SecuritySection/index.tsx
@@ -205,7 +205,7 @@ export function SecuritySection() {
       <AppAutoLockItem />
       <PasswordItem />
       {!platformEnv.isWebDappMode ? <ConnectedSitesItem /> : null}
-      {platformEnv.isExtension ? <FloatingIcon /> : null}
+      {!platformEnv.isExtension ? <FloatingIcon /> : null}
       <SignatureRecordItem />
       <ProtectionItem />
       <CleanDataItem />
